# specs/dao-api-spec.yaml
{
  "paths": {
...
    "/users/{user_id}": {
      "get": {
        "responses": {
          "200": {
            "description": "Get one user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "404": {
            "description": "NotFound",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagingError"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/DEFAULT_ERROR"
          }
        },
        "summary": "Get an existing user",
        "tags": [
          "users"
        ],
        "operationId": "GetUser"
      },
      "put": {
        "responses": {
          "422": {
            "$ref": "#/components/responses/UNPROCESSABLE_ENTITY"
          },
          "200": {
            "description": "Update one user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "404": {
            "description": "NotFound",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagingError"
                }
              }
            }
          },
          "400": {
            "description": "BadRequest",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagingError"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/DEFAULT_ERROR"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InputUpdateUser"
              }
            }
          }
        },
        "summary": "Update an existing user",
        "tags": [
          "users"
        ],
        "operationId": "UpdateUser"
      },
      "parameters": [
        {
          "in": "path",
          "name": "user_id",
          "required": true,
          "schema": {
            "type": "integer",
            "minimum": 0
          }
        }
      ]
    },
...
  },
  "info": {
    "title": "DAO-API",
    "version": "dev"
  },
...
  "openapi": "3.0.2",
  "components": {
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "description": "Error code"
          },
          "status": {
            "type": "string",
            "description": "Error name"
          },
          "message": {
            "type": "string",
            "description": "Error message"
          },
          "errors": {
            "type": "object",
            "description": "Errors",
            "additionalProperties": {}
          }
        }
      },
...
      "PagingError": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "description": "Error status code"
          },
          "message": {
            "type": "string",
            "description": "Error message"
          },
          "status": {
            "type": "string",
            "description": "Error status"
          }
        },
        "description": "Informations about the error"
      },
...
      "User": {
        "type": "object",
        "properties": {
          "user_id": {
            "type": "integer",
            "description": "Unique user identifier"
          },
          "username": {
            "type": "string",
            "description": "Username of the user"
          },
          "email": {
            "type": "string",
            "description": "Email of the user"
          },
          "discord_username": {
            "type": "string",
            "description": "Discord username of the user"
          },
          "wallet_address": {
            "type": "string",
            "description": "Wallet address of the user"
          }
        },
        "description": "User informations."
      },
      "UserResponse": {
        "type": "object",
        "properties": {
          "action": {
            "type": "string"
          },
          "user": {
            "$ref": "#/components/schemas/User"
          }
        },
        "description": "Create/Update a user."
      },
      "InputUpdateUser": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "description": "New Username of the user"
          },
          "email": {
            "type": "string",
            "description": "New Email of the user"
          },
          "password": {
            "type": "string",
            "description": "New Password of the user"
          },
          "discord_username": {
            "type": "string",
            "description": "New Discord username of the user"
          }
        },
        "description": "New user information"
      },
...
    },
    "responses": {
      "UNPROCESSABLE_ENTITY": {
        "description": "Unprocessable Entity",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "DEFAULT_ERROR": {
        "description": "Default error response",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      }
    }
  }
}